backend/LLM/requirements.txt
fastapi
uvicorn
langchain
langchain-groq
langgraph
mcp
python-dotenv
pydantic
